requirements.txt
pbr!=2.1.0  # Apache-2.0
ib3
irc
Jinja2  # BSD License (3 clause)
jsonschema  # MIT
requests!=2.20.0  # Apache-2.0
PyYAML  # MIT